requirements.txt
streamlit
langchain>=0.2.0
langchain-core
langchain-groq
feedparser
python-dotenv
joblib
